Refactor arrays containing fields
